Update Test
Added testName field and a getter.
Commented out the int returning createTest method.
Added updates to the numberOfQuestions in test each time the questions were edited.
Added remove question and topic methods.
Changed parameter list for createTest to take an arrayList of questions rather than an array for the sake of unifomity